Add pushDocker and update microService , loadBalancer and ServerRegistry
Update microService.js in MicroServices

Update the microService.js file in the MicroServices directory to remove unused dependencies and code related to registering the microservice with the discovery server.

Add pushDocker.js to Monitor middleware

Add a new file, pushDocker.js, to the Monitor middleware.This functionality will be used to deploy the Monitor middleware to a remote server.